+++ apps/block_scout_web/lib/block_scout_web/controllers/search_controller.ex
@@ -3,7 +3,8 @@ defmodule BlockScoutWeb.SearchController do
 
   import BlockScoutWeb.Chain, only: [paging_options: 1, next_page_params: 3, split_list_by_page: 1]
 
-  alias BlockScoutWeb.{ChainController, Controller, SearchView}
+  alias BlockScoutWeb.{Controller, SearchView}
+  alias Explorer.Chain
   alias Phoenix.View
 
   def search_results(conn, %{"q" => query, "type" => "JSON"} = params) do
@@ -12,7 +13,7 @@ defmodule BlockScoutWeb.SearchController do
 
     search_results_plus_one =
       paging_options
-      |> ChainController.search_by(offset, query)
+      |> Chain.joint_search(offset, query)
 
     {search_results, next_page} = split_list_by_page(search_results_plus_one)
 
@@ -72,22 +73,5 @@ defmodule BlockScoutWeb.SearchController do
       query: nil,
       current_path: Controller.current_full_path(conn)
     )
-  end
-
-  def api_search_result(conn, %{"q" => query} = params) do
-    [paging_options: paging_options] = paging_options(params)
-    offset = (max(paging_options.page_number, 1) - 1) * paging_options.page_size
-
-    search_results_plus_one =
-      paging_options
-      |> ChainController.search_by(offset, query)
-
-    {search_results, next_page} = split_list_by_page(search_results_plus_one)
-
-    next_page_params = next_page_params(next_page, search_results, params)
-
-    conn
-    |> put_status(200)
-    |> render(:search_results, %{search_results: search_results, next_page_params: next_page_params})
   end
 end
